requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
joblib==1.4.0
Markdown==3.6
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.4.0
tzdata==2024.1
